pyproject.toml
[project]
name = "fastapinextjstemplate1"
version = "0.1.0"
description = ""
authors = [
    {name = "Marcin Marciszewski",email = "marciszewski.m@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.7,<0.116.0)",
    "celery (>=5.4.0,<6.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "alembic (==1.13.1)",
    "jinja2 (>=3.1.5,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "asgiref (>=3.8.1,<4.0.0)",
    "asyncio-redis (>=0.16.0,<0.17.0)",
    "broadcaster (>=0.3.1,<0.4.0)",
    "redis (==5.0.1)",
    "python-socketio (==5.7.1)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pytest-cov (>=6.0.0,<7.0.0)",
    "pytest-factoryboy (>=2.7.0,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
debugpy = "^1.8.12"
watchfiles = "^1.0.4"
pytest = "^8.3.4"
httpx = "^0.28.1"
factory-boy = "^3.3.3"
